2020.08.17
Update

+++ k8s/Misc/create_key.bash
@@ -11,19 +11,26 @@ mv new init.conf
 
 . init.conf
 
+rm key.json > /dev/null 2>&1
 curl --user "${user}:${publicApiKey}" --digest \
   --header "Accept: application/json" \
   --header "Content-Type: application/json" \
-  --request POST "http://opsmgr:8080/api/public/v1.0/admin/apiKeys?pretty=true" \
+  --request POST "${opsMgrUrl}/api/public/v1.0/admin/apiKeys?pretty=true" \
   --data '{
     "desc" : "New API key for Global Testing",
     "roles" : [ "GLOBAL_OWNER" ]
   }' \
   -o key.json > /dev/null 2>&1
 
-cat init.conf |sed -e '/privateKey/d' -e '/publicKey/d' > new
-echo  publicKey="$( cat key.json |jq .username )"
-echo  privateKey="$( cat key.json |jq .privateKey )"
-echo  publicKey="$( cat key.json |jq .username )" >> new
-echo  privateKey="$( cat key.json |jq .privateKey )" >> new
-mv new init.conf
+if [[ -e "key.json" ]]
+then
+    cat init.conf |sed -e '/privateKey/d' -e '/publicKey/d' > new
+    echo  publicKey="$( cat key.json |jq .username )"
+    echo  privateKey="$( cat key.json |jq .privateKey )"
+    echo  publicKey="$( cat key.json |jq .username )" >> new
+    echo  privateKey="$( cat key.json |jq .privateKey )" >> new
+    mv new init.conf
+else
+    printf "%s\n" "Error did not create key"
+    exit 1
+fi
